cleaning self-host project
