+++ SelfHost/SelfHost.csproj
@@ -74,6 +74,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="_OwinHostStuff\HeaderDictionary.cs" />
+    <Compile Include="_OwinHostStuff\TaskEx.cs" />
+    <Compile Include="_OwinHostStuff\SelfHostingApp.cs" />
     <Compile Include="_OwinHostStuff\ContextWrapper.cs" />
     <Compile Include="Index.cs" />
     <Compile Include="Program.cs" />
@@ -99,9 +102,7 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
-    <Content Include="Views\Index.cshtml">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
+    <Content Include="Views\Index.cshtml" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
